Add relative enmity and time to death
Update enmity.js to calculate relative enmity, as that is somewhat
easier to read than absolute enmity over long fights.

Additionally, keep a history of target health to calculate time
until death.

This adds two new html overlays to show this info.  enmity_relative is
an abbreviated enmity list with only relative enmity.  targetinfo_mini
is a short list of numbers intended for being closer to the middle of
the screen.